Feat: 예약 관리 페이지 예약 승인 API
- 관리자 권한에 따라 1차 승인 또는 2차 승인 진행

+++ src/main/java/Team_Mute/back_end/domain/reservation_admin/dto/ReservationListResponseDto.java
@@ -1,12 +1,12 @@
 package Team_Mute.back_end.domain.reservation_admin.dto;
 
-import Team_Mute.back_end.domain.reservation_admin.entity.Reservation;
+import Team_Mute.back_end.domain.reservation.entity.Reservation;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 public class ReservationListResponseDto {
-	public Integer reservationId;
+	public Long reservationId;
 	public String reservationStatusName;
 	public String spaceName;
 	public String userName;
